+++ src/main/java/com/synrc/bert/Enc.java
@@ -3,6 +3,7 @@ package com.synrc.bert;
 import fj.*;
 import static fj.P.p;
 import fj.data.List;
+import java.math.BigDecimal;
 
 public interface Enc<T> extends F<T, Term> {
     Term encode(T v);
@@ -16,7 +17,7 @@ public interface Enc<T> extends F<T, Term> {
     public static Enc<String>   stringEnc = Term::str;
     public static Enc<byte[]>   binEnc  = Term::bin;
     public static Enc<Double>   floatEnc = Term::float754;
-    public static Enc<Double>   floatStrEnc = Term::floatStr;
+    public static Enc<BigDecimal>   floatStrEnc = Term::floatStr;
     public static Enc<Byte>     byteEnc = Term::bt;
     public static Enc<Integer>  intEnc = Term::in;
